--- modules/service/src/main/scala/lucuma/odb/graphql/snippet/ObservationSnippet.scala
@@ -1,380 +1,0 @@
-// Copyright (c) 2016-2022 Association of Universities for Research in Astronomy, Inc. (AURA)
-// For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
-
-package lucuma.odb.graphql
-package snippet
-
-import cats.effect.Sync
-import cats.effect.kernel.Resource
-import cats.syntax.all._
-import edu.gemini.grackle.Cursor
-import edu.gemini.grackle.Cursor.Env
-import edu.gemini.grackle.Path.ListPath
-import edu.gemini.grackle.Path.UniquePath
-import edu.gemini.grackle.Predicate
-import edu.gemini.grackle.Predicate._
-import edu.gemini.grackle.Query
-import edu.gemini.grackle.Query._
-import edu.gemini.grackle.Result
-import edu.gemini.grackle.TypeRef
-import edu.gemini.grackle.skunk.SkunkMapping
-import lucuma.core.enums.CloudExtinction
-import lucuma.core.enums.ImageQuality
-import lucuma.core.enums.SkyBackground
-import lucuma.core.enums.WaterVapor
-import lucuma.core.model.Access._
-import lucuma.core.model.Observation
-import lucuma.core.model.User
-import lucuma.odb.data.Existence
-import lucuma.odb.data.ObsActiveStatus
-import lucuma.odb.data.ObsStatus
-import lucuma.odb.data.UpdateResult
-import lucuma.odb.graphql.snippet.input.CreateObservationInput
-import lucuma.odb.graphql.snippet.input.ObservationPropertiesInput
-import lucuma.odb.graphql.snippet.input.UpdateObservationsInput
-import lucuma.odb.graphql.snippet.input.WhereObservation
-import lucuma.odb.graphql.util.Bindings.BooleanBinding
-import lucuma.odb.graphql.util._
-import lucuma.odb.service.ObservationService
-import lucuma.odb.service.ObservationService.CreateResult.NotAuthorized
-import lucuma.odb.service.ObservationService.CreateResult.Success
-import lucuma.odb.util.Codecs._
-import natchez.Trace
-import skunk.AppliedFragment
-import skunk.Session
-
-object ObservationSnippet {
-
-  def apply[F[_]: Sync: Trace](
-    m:      SnippetMapping[F] with SkunkMapping[F] with MutationCompanionOps[F],
-    dbPool: Resource[F, Session[F]],
-    user:   User
-  ): m.Snippet = {
-
-    import m.ColumnRef
-    import m.Join
-    import m.LeafMapping
-    import m.MappedQuery
-    import m.Mutation
-    import m.MutationCompanionOps
-    import m.ObjectMapping
-    import m.Snippet
-    import m.SqlField
-    import m.SqlObject
-    import m.SqlRoot
-    import m.TableDef
-    import m.col
-    import m.schema
-
-    // The types that we're going to map.
-    val QueryType             = schema.ref("Query")
-    val MutationType          = schema.ref("Mutation")
-    val ObservationType       = schema.ref("Observation")
-    val ObservationIdType     = schema.ref("ObservationId")
-    val ObsStatusType         = schema.ref("ObsStatus")
-    val ObsActiveStatusType   = schema.ref("ObsActiveStatus")
-
-    val TargetEnvironmentType = schema.ref("TargetEnvironment")
-    val CoordinatesType       = schema.ref("Coordinates")
-
-    val ConstraintSetType     = schema.ref("ConstraintSet")
-    val CloudExtinctionType   = schema.ref("CloudExtinction")
-    val ImageQualityType      = schema.ref("ImageQuality")
-    val SkyBackgroundType     = schema.ref("SkyBackground")
-    val WaterVaporType        = schema.ref("WaterVapor")
-    val ElevationRangeType    = schema.ref("ElevationRange")
-    val AirMassRangeType      = schema.ref("AirMassRange")
-    val HourAngleRangeType    = schema.ref("HourAngleRange")
-
-    val CreateObservationResultType  = schema.ref("CreateObservationResult")
-
-    val pool = dbPool.map(ObservationService.fromSessionAndUser(_, user))
-
-    // TODO: Can we share the common predicates somewhere?
-    object Predicates {
-
-      def includeDeleted(b: Boolean): Predicate =
-        if (b) True else Eql(UniquePath(List("existence")), Const[Existence](Existence.Present))
-
-      def hasObservationId(oid: Observation.Id): Predicate =
-        Eql(UniquePath(List("id")), Const(oid))
-
-      def inObservationIds(oids: List[Observation.Id]): Predicate =
-        In(UniquePath(List("id")), oids)
-
-      def isVisibleTo(user: User): Predicate =
-        user.role.access match {
-          case Guest | Pi =>
-            Or(
-              Contains(ListPath(List("program", "users", "userId")), Const(user.id)), // user is linked, or
-              Eql(UniquePath(List("program", "piUserId")), Const(user.id))            // user is the PI
-            )
-          case Ngo => ???
-          case Staff | Admin | Service => True
-        }
-
-      def isWritableBy(user: User): Predicate =
-        isVisibleTo(user)
-
-    }
-
-    // Column references for our mapping.
-    object ObservationView extends TableDef("v_observation") {
-      val ProgramId: ColumnRef    = col("c_program_id",          program_id)
-      val Id: ColumnRef           = col("c_observation_id",      observation_id)
-      val Existence: ColumnRef    = col("c_existence",           existence)
-      val Subtitle: ColumnRef     = col("c_subtitle",            text_nonempty.opt)
-//      val Instrument: m.ColumnRef   = col("c_instrument", tag.opt)
-      val Status: ColumnRef       = col("c_status",              obs_status)
-      val ActiveStatus: ColumnRef = col("c_active_status",       obs_active_status)
-
-      object TargetEnvironment {
-        object Coordinates {
-          val SyntheticId: ColumnRef = col("c_explicit_base_id",  observation_id.embedded)
-        }
-      }
-
-      object ConstraintSet {
-        val CloudExtinction: ColumnRef = col("c_cloud_extinction", cloud_extinction.embedded)
-        val ImageQuality: ColumnRef    = col("c_image_quality",    image_quality.embedded)
-        val SkyBackground: ColumnRef   = col("c_sky_background",   sky_background.embedded)
-        val WaterVapor: ColumnRef      = col("c_water_vapor",      water_vapor.embedded)
-        object ElevationRange {
-          object AirMassRange {
-            val SyntheticId: ColumnRef = col("c_air_mass_id",  observation_id.embedded)
-            val AirMassMin: ColumnRef  = col("c_air_mass_min", air_mass_range_value.embedded)
-            val AirMassMax: ColumnRef  = col("c_air_mass_max", air_mass_range_value.embedded)
-          }
-          object HourAngleRange {
-            val SyntheticId: ColumnRef  = col("c_hour_angle_id",  observation_id.embedded)
-            val HourAngleMin: ColumnRef = col("c_hour_angle_min", hour_angle_range_value.embedded)
-            val HourAngleMax: ColumnRef = col("c_hour_angle_max", hour_angle_range_value.embedded)
-          }
-        }
-      }
-    }
-
-    // Column references for our mapping.
-    object ProgramTable extends TableDef("t_program") {
-      val Id: ColumnRef = col("c_program_id", program_id)
-    }
-
-    // TBD
-//    object AsterismTargetTable extends TableDef("t_asterism_target") {
-//      val ProgramId: ColumnRef     = col("c_program_id",     program_id)
-//      val ObservationId: ColumnRef = col("c_observation_id", observation_id)
-//      val TargetId: ColumnRef      = col("c_target_id",      target_id)
-//    }
-
-    def uniqueObservationNoFiltering(id: Observation.Id, child: Query): Query =
-      Unique(Filter(Predicates.hasObservationId(id), child))
-
-    def observationListNoFiltering(ids: List[Observation.Id], child: Query): Query =
-      Filter(Predicates.inObservationIds(ids), child)
-
-    val createObservation: Mutation =
-      Mutation.simple { (child, env) =>
-        env.getR[CreateObservationInput]("input").flatTraverse { input =>
-          pool.use { svc =>
-            svc.createObservation(input.programId, input.SET.getOrElse(ObservationPropertiesInput.DefaultCreate)).map {
-              case NotAuthorized(user) => Result.failure(s"User ${user.id} is not authorized to perform this action")
-              case Success(id)         => Result(uniqueObservationNoFiltering(id, child))
-            }
-          }
-        }
-      }
-
-    val updateObservations: Mutation =
-      Mutation.simple { (child, env) =>
-        env.getR[UpdateObservationsInput]("input").flatTraverse { input =>
-
-          // Predicate for selecting programs to update
-          val filterPredicate: Predicate = and(List(
-            Predicates.isWritableBy(user),
-            Predicates.includeDeleted(input.includeDeleted.getOrElse(false)),
-            input.WHERE.getOrElse(True)
-          ))
-
-          // An applied fragment that selects all program ids that satisfy
-          // `filterPredicate`
-          val idSelect: Result[AppliedFragment] =
-            Result.fromOption(
-              MappedQuery(
-                Filter(filterPredicate, Select("id", Nil, Query.Empty)),
-                Cursor.Context(ObservationType)
-              ).map(_.fragment),
-              "Could not construct a subquery for the provided WHERE condition."
-            )
-
-          idSelect.traverse { which =>
-            pool.use { svc =>
-              svc
-                .updateObservations(input.SET, which)
-                .map { lst =>
-                  Filter(Predicates.inObservationIds(lst), child)
-                }
-            }
-          }
-        }
-      }
-
-    val typeMappings =
-      List(
-        ObjectMapping(
-          tpe = ObservationType,
-          fieldMappings = List(
-            SqlField("id", ObservationView.Id, key = true),
-            SqlField("programId", ObservationView.ProgramId, hidden=true),
-            SqlField("existence", ObservationView.Existence, hidden = true),
-            SqlField("subtitle", ObservationView.Subtitle),
-            SqlField("status", ObservationView.Status),
-            SqlField("activeStatus", ObservationView.ActiveStatus),
-            SqlObject("targetEnvironment"),
-            SqlObject("constraintSet"),
-            SqlObject("program", Join(ObservationView.ProgramId, ProgramTable.Id))
-          ),
-        ),
-        ObjectMapping(
-          tpe = TargetEnvironmentType,
-          fieldMappings = List(
-            SqlField("id", ObservationView.Id, key = true, hidden = true),
-            SqlObject("explicitBase")
-          )
-        ),
-        ObjectMapping(
-          tpe = CoordinatesType,
-          fieldMappings = List(
-            SqlField("synthetic_id", ObservationView.TargetEnvironment.Coordinates.SyntheticId, key = true, hidden = true),
-            SqlObject("ra"),
-            SqlObject("dec")
-          )
-        ),
-        ObjectMapping(
-          tpe = ConstraintSetType,
-          fieldMappings = List(
-            SqlField("id", ObservationView.Id, key = true, hidden = true),
-            SqlField("cloudExtinction", ObservationView.ConstraintSet.CloudExtinction),
-            SqlField("imageQuality",    ObservationView.ConstraintSet.ImageQuality),
-            SqlField("skyBackground",   ObservationView.ConstraintSet.SkyBackground),
-            SqlField("waterVapor",      ObservationView.ConstraintSet.WaterVapor),
-            SqlObject("elevationRange")
-          )
-        ),
-        ObjectMapping(
-          tpe = ElevationRangeType,
-          fieldMappings = List(
-            SqlField("id", ObservationView.Id, key = true, hidden = true),
-            SqlObject("airMass"),
-            SqlObject("hourAngle")
-          )
-        ),
-        ObjectMapping(
-          tpe = AirMassRangeType,
-          fieldMappings = List(
-            SqlField("synthetic_id", ObservationView.ConstraintSet.ElevationRange.AirMassRange.SyntheticId, key = true, hidden = true),
-            SqlField("min", ObservationView.ConstraintSet.ElevationRange.AirMassRange.AirMassMin),
-            SqlField("max", ObservationView.ConstraintSet.ElevationRange.AirMassRange.AirMassMax)
-          )
-        ),
-        ObjectMapping(
-          tpe = HourAngleRangeType,
-          fieldMappings = List(
-            SqlField("synthetic_id", ObservationView.ConstraintSet.ElevationRange.HourAngleRange.SyntheticId, key = true, hidden = true),
-            SqlField("minHours", ObservationView.ConstraintSet.ElevationRange.HourAngleRange.HourAngleMin),
-            SqlField("maxHours", ObservationView.ConstraintSet.ElevationRange.HourAngleRange.HourAngleMax)
-          )
-        ),
-        ObjectMapping(
-          tpe = MutationType,
-          fieldMappings = List(
-            SqlRoot("createObservation", mutation = createObservation),
-            SqlRoot("updateObservations", mutation = updateObservations)
-          )
-        ),
-        ObjectMapping(
-          tpe = CreateObservationResultType,
-          fieldMappings = List(
-            SqlField("id", ObservationView.Id, key = true, hidden = true),
-            SqlObject("observation"),
-          )
-        ),
-        ObjectMapping(
-          tpe = QueryType,
-          fieldMappings = List(
-            SqlRoot("observation"),
-            SqlRoot("observations")
-          )
-        ),
-        LeafMapping[Observation.Id](ObservationIdType),
-        LeafMapping[ObsStatus](ObsStatusType),
-        LeafMapping[ObsActiveStatus](ObsActiveStatusType),
-        LeafMapping[CloudExtinction](CloudExtinctionType),
-        LeafMapping[ImageQuality](ImageQualityType),
-        LeafMapping[SkyBackground](SkyBackgroundType),
-        LeafMapping[WaterVapor](WaterVaporType)
-      )
-
-    val elaborator = Map[TypeRef, PartialFunction[Select, Result[Query]]](
-      QueryType -> {
-        case Select("observation", List(
-          ObservationIdBinding("observationId", rOid)
-        ), child) =>
-          rOid.map { oid =>
-            Select("observation", Nil,
-              Unique(
-                Filter(
-                  And(
-                    Predicates.hasObservationId(oid),
-                    Predicates.isVisibleTo(user)
-                  ),
-                  child
-                )
-              )
-            )
-          }
-
-        case Select("observations", List(
-          WhereObservation.Binding.Option("WHERE", rWHERE),
-          ObservationIdBinding.Option("OFFSET", rOFFSET),
-          NonNegIntBinding.Option("LIMIT", rLIMIT),
-          BooleanBinding("includeDeleted", rIncludeDeleted)
-        ), child) =>
-          (rWHERE, rOFFSET, rLIMIT, rIncludeDeleted).parMapN { (WHERE, OFFSET, LIMIT, includeDeleted) =>
-            Select("observations", Nil,
-              Limit(
-                LIMIT.foldLeft(1000)(_ min _.value),  // TODO: we need a common place for the max limit
-                Filter(
-                  And.all(
-                    OFFSET.map(oid => GtEql(UniquePath(List("id")), Const(oid))).getOrElse(True),
-                    Predicates.includeDeleted(includeDeleted),
-                    Predicates.isVisibleTo(user),
-                    WHERE.getOrElse(True)
-                  ),
-                  child
-                )
-              )
-            )
-          }
-
-      },
-
-      MutationType -> {
-        case Select("createObservation", List(CreateObservationInput.Binding("input", rInput)), child) =>
-          rInput.map { input =>
-            Environment(Env("input" -> input), Select("createObservation", Nil, child))
-          }
-
-        case Select("updateObservations", List(UpdateObservationsInput.Binding("input", rInput)), child) =>
-          rInput.map { input =>
-            Environment(Env("input" -> input), Select("updateObservations", Nil, child))
-          }
-      }
-    )
-
-    // Done.
-    Snippet(typeMappings, elaborator)
-
-  }
-
-
-}
