+++ modules/service/src/main/scala/lucuma/odb/graphql/snippet/mapping/CreateObservationResultMapping.scala
@@ -1,0 +1,27 @@
+// Copyright (c) 2016-2022 Association of Universities for Research in Astronomy, Inc. (AURA)
+// For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
+
+package lucuma.odb.graphql
+package snippet
+package mapping
+
+import table.ObservationView
+
+import edu.gemini.grackle.skunk.SkunkMapping
+
+trait CreateObservationResultMapping[F[_]]
+  extends ObservationView[F] { this: SkunkMapping[F] =>
+
+  lazy val CreateObservationResultType = schema.ref("CreateObservationResult")
+
+  lazy val CreateObservationResultMapping =
+    ObjectMapping(
+      tpe = CreateObservationResultType,
+      fieldMappings = List(
+        SqlField("id", ObservationView.Id, key = true, hidden = true),
+        SqlObject("observation"),
+      )
+    )
+
+}
+
