+++ modules/service/src/main/scala/lucuma/odb/graphql/snippet/mapping/ElevationRangeMapping.scala
@@ -1,0 +1,28 @@
+// Copyright (c) 2016-2022 Association of Universities for Research in Astronomy, Inc. (AURA)
+// For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
+
+package lucuma.odb.graphql
+package snippet
+package mapping
+
+import table.ObservationView
+
+import edu.gemini.grackle.skunk.SkunkMapping
+
+trait ElevationRangeMapping[F[_]]
+  extends ObservationView[F] { this: SkunkMapping[F] =>
+
+  lazy val ElevationRangeType = schema.ref("ElevationRange")
+
+  lazy val ElevationRangeMapping =
+    ObjectMapping(
+      tpe = ElevationRangeType,
+      fieldMappings = List(
+        SqlField("id", ObservationView.Id, key = true, hidden = true),
+        SqlObject("airMass"),
+        SqlObject("hourAngle")
+      )
+    )
+
+}
+
