+++ modules/service/src/main/scala/lucuma/odb/graphql/snippet/mapping/TargetEnvironmentMapping.scala
@@ -1,0 +1,26 @@
+// Copyright (c) 2016-2022 Association of Universities for Research in Astronomy, Inc. (AURA)
+// For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
+
+package lucuma.odb.graphql
+package snippet
+package mapping
+
+import table.ObservationView
+import edu.gemini.grackle.skunk.SkunkMapping
+
+trait TargetEnvironmentMapping[F[_]]
+  extends ObservationView[F] { this: SkunkMapping[F] =>
+
+  lazy val TargetEnvironmentType = schema.ref("TargetEnvironment")
+
+  lazy val TargetEnvironmentMapping =
+    ObjectMapping(
+      tpe = TargetEnvironmentType,
+      fieldMappings = List(
+        SqlField("id", ObservationView.Id, key = true, hidden = true),
+        SqlObject("explicitBase")
+      )
+    )
+
+}
+
